requirements.txt
oauth2client==4.1.3
google-auth-oauthlib==1.2.1
gspread==6.0.2
gspread-dataframe==3.3.1
openpyxl==3.1.5
pandas==2.1.4
pyngrok==7.2.0
requests==2.31.0
requests-oauthlib==1.3.1
Flask==2.2.5
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.27.0
google-generativeai==0.7.2
google-auth-httplib2==0.2.0
pdf2image==1.17.0
Pillow==9.4.0
pytesseract==0.3.10